+++ CodeConverter/CSharp/VBToCSConversion.cs
@@ -48,11 +48,11 @@ public class VBToCSConversion : ILanguageConversion
         if (!string.IsNullOrEmpty(rootNamespaceDot)) rootNamespaceDot += ".";
 
         return new[] {
-            ("\\\\Microsoft\\.VisualBasic\\.targets", "\\Microsoft.CSharp.targets"),
-            ("\\.vb\"", ".cs\""),
-            ("\\.vb<", ".cs<"),
-            ("<\\s*Generator\\s*>\\s*VbMyResourcesResXFileCodeGenerator\\s*</\\s*Generator\\s*>", "<Generator>ResXFileCodeGenerator</Generator>"),
-            ("(<\\s*CustomToolNamespace\\s*>)(.*</\\s*CustomToolNamespace\\s*>)", $"$1{rootNamespaceDot}$2"), // <CustomToolNamespace>My.Resources</CustomToolNamespace>
+            (@"\\([A-Za-z\.]*)\.VisualBasic\.targets", @"\$1.CSharp.targets"),
+            (@"\.vb""", @".cs"""),
+            (@"\.vb<", @".cs<"),
+            (@"<\s*Generator\s*>\s*VbMyResourcesResXFileCodeGenerator\s*</\s*Generator\s*>", @"<Generator>ResXFileCodeGenerator</Generator>"),
+            (@"(<\s*CustomToolNamespace\s*>)(.*</\s*CustomToolNamespace\s*>)", @$"$1{rootNamespaceDot}$2"), // <CustomToolNamespace>My.Resources</CustomToolNamespace>
         };
     }